Block setting BottomSheet internal state and hide if content null
This patch blocks an invalid state and silently reports a stack to
chrome's crash reporter which will allow us to continue to track
down the root cause of the issue. A side-effect of adding this logic
allowed us to track down another long-standing issue with the
"settle animation" where the internal state would be set to "none".
This was an issue with operation order: canceling animation after
the new state was set.


Bug: 1126872
Change-Id: I0dae62be822d7f3fe8e116597d17b08899b3e239
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2416871
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808563}