Speculative fix for InstallableUtils crash
The WebAppProvider can be uninitialized. Dependencies should use the
on_registry_ready() signal. This change is a quick & mergeable fix for
a related crash, after which the codepaths will be turned async if
possible.

TBR-ing dominick so this change can get a canary roll, and they can
look at it when they get in on Monday. Patch is very low-risk -
resolving crash when there used to be a crash.

R=pwnall@chromium.org

TBR: dominickn@chromium.org
Bug: 1129921
Change-Id: I57c6d4f181f28c178df85c60626423a5619f9494
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2419537
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808617}