Tab Search: Add fixed Tab Search entrypoint to the TabStripRegionView
This CL partially reverts the following CL's initial implementation
of the Tab Search fixed entrypoint to address functional conflicts
with the Tab Scrolling feature.
https://crrev.com/c/2377887

When the kTabSearchFixedEntrypoint flag is enabled the
TabSearchButton will now appear as a child of the TabStripRegionView.
This allows the the TabSearchButton to function as intended whilst
unblocking changes needed for Tab Scrolling.

This CL also updates the TabStripRegionView such that several
existing method calls made to the TabStrip now pass through the
TabStripRegionView. This makes the implementation of the fixed
entrypoint saner and is a step in the right direction wrt existing
efforts to move the NTB out of the TabStrip and into the
TabStripRegionView.

Bug: 1099917
Change-Id: I5b73c5cc21a7b90086639e68efa8fc79a40a2beb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2416681
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808658}