[macOS Capture] On cameras with broken MJPEG, prefer NV12 instead.
We have avoided capturing MJPEG due to a known issue with a specific
camera model.
In the event of MJPEG being picked (e.g. doing 4K on this camera), we
fall back on UYVY (2vuy) instead, which used to be our default pixel
format.

Thanks to [1] we no longer avoid MJPEG unless the camera model is the
problematic one. This leads to better performance.

However in the event that we are on the problematic model and do want
to avoid MJPEG, requesting UYVY makes less sense than requesting NV12.
In either case macOS has to do a conversion for us, because neither
format is natively supported in the requested resolution. If we're
going to request a non-supported format, we might as well request a
format that we want to be using: NV12, which is the new default
format.

In the future when we can encode NV12, this will avoid an additional
conversion from UYVY to NV12, since macOS did the correct conversion
for us.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2412732

Bug: 1116241, 1124884, 1035077
Change-Id: Ibce938f99194057778ce760ff9b3c4ec68c9b430
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2416289
Reviewed-by: Markus Handell <handellm@google.com>
Commit-Queue: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/master@{#808681}