Adding SAML password sync token to prefs.
SAML password sync token is stored in local settings and accessed by
known_user interface (this is to make it available on the login
screen). As has been pointed out in cl/2359998 this will not work for
ephemeral users. To handle this case we introduce a new property in
prefs. In session we rely only on the pref value while every write
updates both pref and local settings.

Bug: 1090341
Change-Id: I921257ee6904508b6807c7408668e4c2d6f2abbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2410122
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808748}