Update RFH::last_http_status_code in DidCommitNewDocument.
The RFH::last_http_status_code must not change after a same-document
navigation.

The previous code was relying on blink for reporting the correct status
code for every navigation. It was hoping blink won't report a different
status code for same-document navigation.

This patch move setting this value in DidCommitNewDocument(). As a
result it can change only after cross-document navigation.

This is better, but still not ideal. Ideally, we shouldn't depend on the
renderer. We should be able to retrieve this information from the
NavigationRequest instead. This will be done as part of:
https://docs.google.com/document/d/1oUKSDYIaz1-XMJw2f3KRTvq7yNR5O5IK8Iwgmk8laiw/edit#

Bug: None
Change-Id: I50b42f825046ba76eb509d01373ce60d7087dd0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418957
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808835}