Update RFH::last_http_method in DidCommitNewDocument.
The RFH::last_http_method must not change after a same-document
navigation.

The previous code was relying on blink for reporting the correct method
name for every navigation. It was hoping blink won't report a different
method name for same-document navigation.

This patch move setting this value in DidCommitNewDocument(). As a
result it can change only after cross-document navigation.

This is better, but still not ideal. Ideally, we shouldn't depend on the
renderer. We should be able to retrieve this information from the
NavigationRequest instead. This will be done as part of:
https://docs.google.com/document/d/1oUKSDYIaz1-XMJw2f3KRTvq7yNR5O5IK8Iwgmk8laiw/edit#

Bug: None
Change-Id: I66d9f3718db2679f8c8df9d9547e0fcd415fc821
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418676
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808838}