Correctly attribute popups to WebContents
Popups triggered by a different WebContents that then one that actually
opens it were being misattributed to the former.

This is a partial revert of https://crrev.com/c/2277737 where we thought
that source_frame was the frame that opened the popup which is not the
case.

Added a test to prevent future regressions.

Bug: 1128495
Change-Id: I51511716857b0074dad90feee709440c7b5f8795
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418324
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808839}