Extract performance.measureMemory implementation to a new class
Currently performance.measureMemory invokes v8::Isolate::MeasureMemory
directly. This CL decouples that and prepares for the upcoming switch
to the implementation based on PerformanceManager.

The CL is pure refactoring without observable behavour change. It moves
the existing code from Performance::measureMemory and
MeasureMemoryDelegate into a new class called MeasureMemoryController.

The new class is responsible for starting the measurement and resolving
the JavaScript promise upon completion of the measurement.

As a drive-by change this removes the code of MeasureMemoryExperiment
that is no longer running.

Bug: 1085129
Change-Id: I83382405278f8d6013583cd0f6ad79d8e7fc8d8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418661
Commit-Queue: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808842}