Allow multi sign-in even when policy-pushed trusted CA certs are active
Historically, multi sign-in was disallowed when either the primary
Profile or the Profile being added used / had used policy-pushed CA
certificates. The reason for limiting this was that the certificate
verifier used at that time (NSS) had process-wide caches, so certificate
verification results could leak across Profiles.

Now that we use the built-in certificate verifier, this risk does not
exist, so allow multi sign-in regardless of the policy-pushed
certificate state of the primary / secondary Profiles.

Also move clearing of the local_state pref that tracks policy-pushed
certificate usage for a Profile into UserSessionManagerImpl because
MultiProfileUserController does not have a dependency on the
policy-pushed certs infrastructure anymore. In a follow-up the pref can
likely be migrated to be a user Profile pref now instead of a
local_state pref.

Bug: 718002
Test: unit tests, browser tests
Change-Id: I1006e0cb1cb4c9fdd68f5373e18acb27f5c3eaea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297566
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808879}