Reorder and simplify RequestPaint()
Reorder the function to check for a paintserver resource first and then
return any color. This structure should be more conducive to future
paintserver handling.

Change so that we don't resolve a color if there is no color value to
resolve. Also apply the same restriction to the visited paint/color,
such that only a color can "replace" a color.

Bug: 109212
Change-Id: I81a0c47b694406c931663804df6705f7bf651572
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418372
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#808885}