Ozone: Configure Display Controllers of the same DRM only
Split the displays passed to the ScreenManager so that the list of
configs all share the same DRM.
This ensures that all parallel operations are done on the same DRM (i.e.
Atomically committing all display props on the same commit, which
requires the same fd).

BUG=1082882
TEST=no changes in enabling or disabling displays.

Change-Id: I2f38449c2114f71672f409b50808a92d72f758db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2405414
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Mark Yacoub <markyacoub@google.com>
Cr-Commit-Position: refs/heads/master@{#808901}