Add zero-touch enrollment information screens where Alt+V works
This CL inserts the info discussed in go/zt-alt+v-id in the "version"
string which already contains the serial number, while preserving the
existing placement of the OS version on the left and the serial number
on the right.

See screenshots in bug. This CL shows the ZTE indication and always
includes the ADID even when it is the same as the SN).

Bug: chromium:1124875
Test: Build Chrome, deploy and restart UI, see ZTE info if warranted
Change-Id: I2a70098546472b2a237ff6f9a0f82f375ab3f3fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2404689
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808915}