Adjust Zvmul test error threshold for mac-arm64
The threshold for the vector math Zvmul test is a little too tight for
mac-arm64.  Hence update the threshold from 1.5e-5 to 1.9e-5 but only
on mac-arm64.  We keep the threshold the same for other mac platforms
so we can catch any regressions there.

Bug: 1128434
Change-Id: Ib51fa772e27378fb8022e19790ecdfcf1e1a6e7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2419539
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808951}