[CrOS Settings] Fix tablet navigation setting visibility
This fixes two issues with tablet navigation button settings:
(1) This setting is now shown on all devices with tablet capabilities,
    even if tablet mode is not currently active. Previously, it was
    only shown when tablet mode was active.
(2) The settings search result for this toggle is now only shown on
    tablet-capable devices. Previously, it was always shown, even in
    situations where the setting was not available.

This CL also moves the location of the
IsBoardTypeMarkedAsTabletCapable() function to //ash/public where it can
be accessed by settings. Additionally, it removes
accessibility_handler_unittest.cc, since the only thing tested in this
file was the previous functionality which would only show tablet
navigation button settings when tablet mode was active.

Bug: 1111564
Change-Id: Iefa982ab3a28fae6ad40fba827cfd658c09fe98b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2420771
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808952}