[Web App Migration] Add a (failing) test for an uninstalled extension
We can migrate extension-apps to web-apps, but should do so only if
the user still has the extension-app installed (i.e., hasn't explicitly
chosen to remove it previously).

Unfortunately, this currently doesn't work - the preinstalled app will
still be installed. Add a test for this (with the expectations flipped)
so we have it for when it's fixed.

Bug: 1130619
Change-Id: Iba936e0311eb24b4d6b6c7460c1790d895df6d64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2416946
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808968}