WebGPU: Flush eagerly after GPUQueue::submit
In other parts of the WebGPU API, we flush lazily at the end of the
task because the page needs to yield back to the browser for pending
callbacks to resolve anyway. Queue submissions however may be holding
lots of data which the GPU process can begin working on promptly.
Frequent queue submissions will have poor command throughput if we don't
flush eagerly.

In the future, the WebGPU command buffer should have a better
way for the Renderer process to ask the GPU process to begin processing
work, and update the GPU process (in a non-blocking way) of the latest
command it has finished writing. This would yield better throughput
without requiring frequent flushes.

Bug: 1128293
Change-Id: I19903dffc946865b11512d100b711409857b025f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2417639
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809002}