Switch the Linux official build to PGO (from AFDO).
This switches the Linux official build from AFDO to PGO. This yield some
3-4% performance gains on Speedometer and help unifying our desktop
build configurations, PGO is already used on Win and Mac. The PGO
profiles for Linux are generated by a continuous bot and are rolled in
Chrome once or twice per day (AFDO is based on the Android profiles).

This increases binary size by ~3-4%.

Bug: 1111437
Change-Id: I565eb336fa5ad9480e99808ed49246634255b82d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2404620
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809040}