cfm: Prevent joining the same process multiple times
The current paradigm is to give a |CfmServiceContext| to any service
that requests to register itself with it; however, using an invitation
to connect to the same process multiple times is unsupported.

The resolution to this issue is for service_connection to become an
intermediate wrapper for the |CfmServiceContext| inside chromium and
proxy commands to the platform side implementation.

BUG=chromium:1105567, b:165864612, b:160960260
TEST='autoninja -C out/Default chromeos:chromeos_unittests \
     && ./out/Default/chromeos_unittests --dbus-stub \
    --single-process-tests \
    --gtest_filter=CfmServiceConnectionTest.*'

Change-Id: I99848666eb14db1ca7c1160f12900824d7881176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2422137
Commit-Queue: Kyle Williams <kdgwill@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809074}