Move dm_token_utils to policy directory
dm_token_utils.h is used outside of safe_browsing and so targets that
depend on dm_token_utils would need to take a dependency onf
safe_browsing. Instead this CL moves dm_token_utils to the policy build
target, which is likely to already be a dependency of anything that was
going to use it.

Bug: chromium:1078512
Change-Id: I1a61e79e45478fb0c972049dd1df3c4b023f5049
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2388500
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Zach Trudo <zatrudo@google.com>
Cr-Commit-Position: refs/heads/master@{#809089}