Remove fast/multicol/balance-line-overflow.html
This tested some behavior in the legacy fragmentation engine that wasn't
strictly valid, but was a work-around to avoid even more severe
over-stretching during column balancing. In the legacy engine, overflow
affects column sizes, which is generally wrong (but necessary when the
engine doesn't do proper fragmentation in the first place).

The LayoutNG block fragmentation engine failed this test because it just
let the lines overflow the columns (rather than stretching them to fit
the overflow), which is correct behavior.

Bug: 829028
Change-Id: I25142cad7109f028fc783578bdfe7021488cd3e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2421658
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809106}