Android: completely disable testManualInitiationWithBarOpen
No change to logic. This completely disables
testManualInitiationWithBarOpen(). This was previously disabled for
M-only, but this is also flaking on P.

This removes filters in favor of the @DisabledTest annotation since it
appears the test itself is flaky, not just on emulators.

Tbr: dtrainor@chromium.org
Bug: 1036553, 1130712, 1127786
Test: out/Default/bin/run_chrome_public_test_apk -f TranslateCompactInfoBarTest.*
Change-Id: Ic588962d6625a310549d0183754b2d4d612f4643
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2422706
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809120}