Remove a number of high-use metrics from UKM's UseCounter metric.
All of these are above the guidance at the top of ukm_features.cc:
"e.g. <1% of page views" - verified on Windows and Android
stable.

The guidance is there because UKM's sampling is on a per event
basis, so individual enum entries can't be sampled at different
rates. If data in UKM still needs to be collected for these,
they can be added as their own dedicated UKM events and thus
could be sampled appropriately.

Bug: 1129061
Change-Id: Id3087b51eb3af80679026444dc6e4c1aa814d1ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2414917
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Jason Chase <chasej@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809129}