[omnibox] Headers: Fix click through header to bookmark bar on MacOS
We previously added a top inset to section headers to pad it from the
results with no headers.

We didn't consider that the section header could be in the top position.

This leads to some extra margin for no reason, and also leads to being
able to click through the popup onto the bookmark bar on MacOS.

This CL fixes that. We should merge this to M86 as it's a trivial fix.

Bug: 1130812
Change-Id: Ife0e81c343465e6c8624eeecbea7672e9028be93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2422887
Auto-Submit: Tommy Li <tommycli@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809147}