[IntentHandling] Add DumpWithoutCrashing for empty string for preferred apps.
There are couple of cases where when selecting "Stay In Chrome" and
"Remember my choice" actually stores an entry with empty app_id. However
this happens randomly and cannot be reproduced easily. Add the
DumpWithoutCrashing to get a trace to help debugging. The main purpose
is to identify where the empty string get generated, because logically
it doesn't seem to be possible.

In addition to this, some early return are also added to avoid storing
the entries.

BUG=853604

Change-Id: I320f29418b606288e056d38088f266c817f812e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418279
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809171}