Don't store URL if the Tab is not initialized
In EditUrlSuggestionProcessor and EditUrlSuggestionUnitTest
there is an assumption that the URL can be acquired off an
uninitialized Tab. This is currently not allowed using
the UserData pattern. I recommend if the Tab is not
initialized we pull the url off the web contents and not
store on the UserData object CriticalPersistedTabData.

Bug: 1117396
Change-Id: I0223d481b7a25722f4e6bf85e78cad45060fcbd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2410346
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/master@{#809176}